+++ ui/components/ui/entities/get-provider-logo.tsx
@@ -5,6 +5,7 @@ import {
   AzureProviderBadge,
   GCPProviderBadge,
   GitHubProviderBadge,
+  IacProviderBadge,
   KS8ProviderBadge,
   M365ProviderBadge,
 } from "@/components/icons/providers-badge";
@@ -24,6 +25,8 @@ export const getProviderLogo = (provider: ProviderType) => {
       return <M365ProviderBadge width={35} height={35} />;
     case "github":
       return <GitHubProviderBadge width={35} height={35} />;
+    case "iac":
+      return <IacProviderBadge width={35} height={35} />;
     default:
       return null;
   }
@@ -43,6 +46,8 @@ export const getProviderName = (provider: ProviderType): string => {
       return "Microsoft 365";
     case "github":
       return "GitHub";
+    case "iac":
+      return "Infrastructure as Code";
     default:
       return "Unknown Provider";
   }